# backend/api/openapi/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Digital Store API",
    "description": "A digital product e-commerce system API providing endpoints for product browsing, shopping cart management, order processing, payment handling, and admin operations.",
    "version": "1.0",
    "contact": {
      "name": "API Support",
      "email": "support@digitalstore.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Type \"Bearer\" followed by a space and JWT token."
    }
  },
...
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Register a new user",
        "description": "Create a new user account with email, password, and profile information",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/RegisterRequest" }
          }
        ],
        "responses": {
          "201": { "description": "User created successfully", "schema": { "$ref": "#/definitions/UserResponse" } },
          "400": { "description": "Invalid request or email already registered", "schema": { "$ref": "#/definitions/ErrorResponse" } }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "User login",
        "description": "Authenticate user and return JWT access and refresh tokens. Response time < 200ms.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/LoginRequest" }
          }
        ],
        "responses": {
          "200": { "description": "Login successful", "schema": { "$ref": "#/definitions/AuthTokenResponse" } },
          "401": { "description": "Invalid credentials or account locked", "schema": { "$ref": "#/definitions/ErrorResponse" } }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["auth"],
        "summary": "User logout",
        "description": "Invalidate the current access and refresh tokens",
        "security": [{"BearerAuth": []}],
        "produces": ["application/json"],
        "responses": {
          "204": { "description": "Logged out successfully" },
          "401": { "description": "Unauthorized", "schema": { "$ref": "#/definitions/ErrorResponse" } }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["auth"],
        "summary": "Refresh access token",
        "description": "Get a new access token using a valid refresh token",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["refresh_token"],
              "properties": {
                "refresh_token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIs..." }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Token refreshed", "schema": { "$ref": "#/definitions/AuthTokenResponse" } },
          "401": { "description": "Invalid refresh token", "schema": { "$ref": "#/definitions/ErrorResponse" } }
        }
      }
    },
    "/auth/password-reset/request": {
      "post": {
        "tags": ["auth"],
        "summary": "Request password reset",
        "description": "Send a password reset link to the user's email. Link expires after 1 hour.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["email"],
              "properties": {
                "email": { "type": "string", "format": "email", "example": "user@example.com" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Reset email sent (always returns success for security)" }
        }
      }
    },
    "/products": {
      "get": {
        "tags": ["products"],
        "summary": "List products",
        "description": "Get paginated list of active products. Supports filtering by category, price range, and sorting. Response time < 200ms.",
        "produces": ["application/json"],
        "parameters": [
          { "in": "query", "name": "page", "type": "integer", "default": 1 },
          { "in": "query", "name": "per_page", "type": "integer", "default": 20 },
          { "in": "query", "name": "category_id", "type": "integer" },
          { "in": "query", "name": "min_price", "type": "number" },
          { "in": "query", "name": "max_price", "type": "number" },
          { "in": "query", "name": "sort", "type": "string", "enum": ["price_asc", "price_desc", "name_asc", "created_desc"] }
        ],
        "responses": {
          "200": { "description": "Product list with pagination metadata" }
        }
      }
    },
    "/products/search": {
      "get": {
        "tags": ["products"],
        "summary": "Search products",
        "description": "Full-text search on product name and description. Response time < 500ms.",
        "produces": ["application/json"],
        "parameters": [
          { "in": "query", "name": "q", "type": "string", "required": true, "description": "Search query" },
          { "in": "query", "name": "page", "type": "integer", "default": 1 },
          { "in": "query", "name": "per_page", "type": "integer", "default": 20 }
        ],
        "responses": {
          "200": { "description": "Search results with pagination" }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": ["products"],
        "summary": "Get product detail",
        "description": "Get full product information including SKUs, images, and specifications. Response time < 200ms.",
        "produces": ["application/json"],
        "parameters": [
          { "in": "path", "name": "id", "type": "integer", "required": true }
        ],
        "responses": {
          "200": { "description": "Product detail with SKUs" },
          "404": { "description": "Product not found", "schema": { "$ref": "#/definitions/ErrorResponse" } }
        }
      }
    },
...
    "/cart/items": {
      "post": {
        "tags": ["cart"],
        "summary": "Add item to cart",
        "description": "Add a SKU to the shopping cart. Validates inventory availability.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["sku_id", "quantity"],
              "properties": {
                "sku_id": { "type": "integer", "example": 1 },
                "quantity": { "type": "integer", "example": 1, "minimum": 1 }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Item added to cart" },
          "400": { "description": "Invalid request or insufficient inventory" }
        }
      }
    },
    "/orders": {
      "post": {
        "tags": ["orders"],
        "summary": "Create order",
        "description": "Create a new order from cart items. Supports both authenticated users and guests. Response time < 500ms.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/CreateOrderRequest" }
          }
        ],
        "responses": {
          "201": { "description": "Order created" },
          "400": { "description": "Invalid request or empty cart" }
        }
      },
      "get": {
        "tags": ["orders"],
        "summary": "List user orders",
        "description": "Get paginated list of orders for the authenticated user",
        "security": [{"BearerAuth": []}],
        "produces": ["application/json"],
        "responses": {
          "200": { "description": "Order list" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/orders/track": {
      "post": {
        "tags": ["orders"],
        "summary": "Track order",
        "description": "Track an order by order number and email (for guest orders)",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["order_number", "email"],
              "properties": {
                "order_number": { "type": "string", "example": "ORD-20240101-001" },
                "email": { "type": "string", "format": "email", "example": "user@example.com" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Order details" },
          "404": { "description": "Order not found" }
        }
      }
    },
    "/payments/online/session": {
      "post": {
        "tags": ["payments"],
        "summary": "Create online payment session",
        "description": "Create a Stripe checkout session for online payment. Response time < 500ms.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["order_id"],
              "properties": {
                "order_id": { "type": "integer", "example": 1 }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Payment session created with redirect URL" },
          "400": { "description": "Invalid order or already paid" }
        }
      }
    },
    "/payments/webhook/stripe": {
      "post": {
        "tags": ["payments"],
        "summary": "Stripe webhook",
        "description": "Handle Stripe payment webhooks. Verifies signature using HMAC-SHA256.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          { "in": "header", "name": "Stripe-Signature", "type": "string", "required": true }
        ],
        "responses": {
          "200": { "description": "Webhook processed" },
          "400": { "description": "Invalid signature or payload" }
        }
      }
    },
    "/payments/transfer/upload": {
      "post": {
        "tags": ["payments"],
        "summary": "Upload transfer proof",
        "description": "Upload a transfer payment proof image (JPEG, PNG, PDF). Max size: 10MB.",
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "parameters": [
          { "in": "formData", "name": "order_id", "type": "integer", "required": true },
          { "in": "formData", "name": "proof", "type": "file", "required": true }
        ],
        "responses": {
          "200": { "description": "Proof uploaded, order status updated to pending_transfer" },
          "400": { "description": "Invalid file or order" }
        }
      }
    },
    "/admin/login": {
      "post": {
        "tags": ["admin-auth"],
        "summary": "Admin login",
        "description": "Authenticate admin user. Returns token with 8-hour expiration. Requires admin role.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/LoginRequest" }
          }
        ],
        "responses": {
          "200": { "description": "Admin login successful" },
          "401": { "description": "Invalid credentials or insufficient permissions" }
        }
      }
    }
  },
  "definitions": {
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "success": { "type": "boolean", "example": false },
        "error": {
          "type": "object",
          "properties": {
            "code": { "type": "string", "example": "BAD_REQUEST" },
            "message": { "type": "string", "example": "Invalid request parameters" },
            "details": { "type": "string", "example": "Field 'email' is required" }
          }
        },
        "request_id": { "type": "string", "example": "abc123xyz456" }
      }
    },
    "RegisterRequest": {
      "type": "object",
      "required": ["email", "password", "full_name"],
      "properties": {
        "email": { "type": "string", "format": "email", "example": "user@example.com" },
        "password": { "type": "string", "minLength": 8, "maxLength": 72, "example": "SecurePass123!" },
        "full_name": { "type": "string", "minLength": 2, "maxLength": 100, "example": "John Doe" },
        "phone": { "type": "string", "example": "+1234567890" }
      }
    },
    "LoginRequest": {
      "type": "object",
      "required": ["email", "password"],
      "properties": {
        "email": { "type": "string", "format": "email", "example": "user@example.com" },
        "password": { "type": "string", "example": "SecurePass123!" }
      }
    },
    "UserResponse": {
      "type": "object",
      "properties": {
        "success": { "type": "boolean", "example": true },
        "data": {
          "type": "object",
          "properties": {
            "id": { "type": "integer", "example": 1 },
            "email": { "type": "string", "example": "user@example.com" },
            "full_name": { "type": "string", "example": "John Doe" },
            "phone": { "type": "string", "example": "+1234567890" },
            "role": { "type": "string", "example": "user" },
            "is_active": { "type": "boolean", "example": true }
          }
        }
      }
    },
    "AuthTokenResponse": {
      "type": "object",
      "properties": {
        "success": { "type": "boolean", "example": true },
        "data": {
          "type": "object",
          "properties": {
            "access_token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIs..." },
            "refresh_token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIs..." },
            "token_type": { "type": "string", "example": "Bearer" },
            "expires_in": { "type": "integer", "example": 86400, "description": "Token expiration in seconds" }
          }
        }
      }
    },
    "CreateOrderRequest": {
      "type": "object",
      "required": ["shipping_address", "payment_method"],
      "properties": {
        "shipping_address": { "type": "string", "example": "123 Main St, City, Country" },
        "payment_method": { "type": "string", "enum": ["online", "transfer"], "example": "online" },
        "guest_email": { "type": "string", "format": "email", "example": "guest@example.com" },
        "guest_name": { "type": "string", "example": "Guest User" },
        "guest_phone": { "type": "string", "example": "+1234567890" }
      }
    },
...
  }
}